falta por fijar

+++ u25-Ejercicio1/src/main/java/com/team02/u25/ejercicio1/dao/IFabricanteDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.team02.u25.ejercicio1.dto.Fabricante;
 
-public interface IFabricanteDAO extends JpaRepository<Fabricante, Long>{
+public interface IFabricanteDAO extends JpaRepository<Fabricante, Integer>{
 
 }
